fix: rename math to avoid conflict with system math module
